server/requirements.txt
fastapi==0.117.1
python-dotenv==1.1.1
annotated-types==0.7.0 
anyio==4.11.0 
cachetools==5.5.2 
certifi==2025.8.3 
charset-normalizer==3.4.3 
click==8.3.0
google-api-core==2.25.1 
google-api-python-client==2.183.0 
google-auth==2.40.3 
google-auth-httplib2==0.2.0 
google-auth-oauthlib==1.2.2 
googleapis-common-protos==1.70.0 
h11==0.16.0 
httplib2==0.31.0 
idna==3.10 
oauthlib==3.3.1 
proto-plus==1.26.1 
protobuf==6.32.1 
pyasn1==0.6.1 
pyasn1_modules==0.4.2 
pydantic==2.11.9 
pydantic_core==2.33.2 
pyparsing==3.2.5 
requests==2.32.5 
requests-oauthlib==2.0.0 
rsa==4.9.1 
sniffio==1.3.1 
starlette==0.48.0 
typing-inspection==0.4.1 
typing_extensions==4.15.0 
uritemplate==4.2.0 
urllib3==2.5.0 
uvicorn==0.37.0
